+++ decidim-bulletin_board-app/app/models/pending_message.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class PendingMessage < ApplicationRecord
+  include Message
+
   belongs_to :client
   belongs_to :election
